fix fomratting

+++ src/i2p/tun/__main__.py
@@ -46,7 +46,7 @@ class Handler(i2cp.I2CPHandler):
         self._scr.addstr(2, 1, "dst: {}".format(self._dest))
         self._scr.addstr(4, 1, "write buff: {}".format('#' * len(self._write_buff)))
         self._scr.addstr(5, 1, "read buff:  {}".format('#' * len(self._read_buff)))
-        self._scr.addstr(7, 1, "link speed: {} KBps".format(self._bw / 1024))
+        self._scr.addstr(7, 1, "link speed: {} KBps".format(int(self._bw / 1024)))
         self._scr.refresh()
         self._bw = 0
         self.loop.call_later(1, self.update_ui)
